LayoutTests/editing/editability/set-invalid-value-expected.txt

When contentEditable is set to invalid string, SYNTAX_ERR exception (code: 12) must be raised.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".





FAIL exceptionThrown should be true. Was false.
FIXME: setContentEditable with invalid string should throw exception. https://bugs.webkit.org/show_bug.cgi?id=52057
PASS document.getElementById("div").getAttribute("contentEditable") is "abc"
FAIL document.getElementById("div").contentEditable should be true. Was inherit.
PASS document.getElementById("div").isContentEditable is true
FAIL window.getComputedStyle(div, "").getPropertyValue("-webkit-user-modify") should be read-write. Was read-only.

PASS successfullyParsed is true

TEST COMPLETE
